Setup most of the api vs service. Need to decide on db model

+++ glicko/src/main/java/solutions/desperate/glicko/domain/service/glicko/Glicko.java
@@ -4,6 +4,6 @@ import solutions.desperate.glicko.domain.model.Player;
 
 public interface Glicko {
     Player defaultPlayer(String name);
-    Player glicko2(Player player1, Player player2, int result);
+    Player glicko2(Player player1, Player player2, double result);
     Player noGamesRd(Player player);
 }
